fix/profile components separate from pages
